+++ lib/project_types/rails/commands/create.rb
@@ -55,7 +55,7 @@ module Rails
           scopes: 'write_products,write_customers,write_draft_orders',
         ).write(@ctx)
 
-        partners_url = "https://partners.shopify.com/#{form.organization_id}/apps/#{api_client['id']}"
+        partners_url = "#{partners_endpoint}/#{form.organization_id}/apps/#{api_client['id']}"
 
         @ctx.puts(@ctx.message('rails.create.info.created', form.title, partners_url))
         @ctx.puts(@ctx.message('rails.create.info.serve', form.name, ShopifyCli::TOOL_NAME))
@@ -171,6 +171,11 @@ module Rails
 
       def install_gem(name, version = nil)
         Gem.install(@ctx, name, version)
+      end
+
+      def partners_endpoint
+        return 'https://partners.myshopify.io' if @ctx.getenv(ShopifyCli::PartnersAPI::LOCAL_DEBUG)
+        'https://partners.shopify.com'
       end
     end
   end