+++ lib/shopify-cli/messages/messages.rb
@@ -177,6 +177,7 @@ module ShopifyCli
         api: {
           error: {
             internal_server_error: '{{red:{{x}} An unexpected error occurred on Shopify.}}',
+            internal_server_error_debug: "\n{{red:Response details:}}\n%s\n\n",
           },
         },