+++ test/project_types/node/commands/create_test.rb
@@ -235,10 +235,6 @@ module Node
 
       private
 
-      def expect_command(command, chdir: @context.root)
-        @context.expects(:system).with(*command, chdir: chdir)
-      end
-
       def perform_command
         run_cmd("create node \
           --name=test-app \